+++ src/uml/diagrams/sequence/messages/Message.java
@@ -1,4 +1,4 @@
-package uml.diagrams.sequence.message;
+package uml.diagrams.sequence.messages;
 
 import uml.diagrams.sequence.exceptions.*;
 import uml.diagrams.sequence.lifelines.Lifeline;
