+++ src/uml/diagrams/sequence/sequencediagrams/SequenceDiagram.java
@@ -1,12 +1,11 @@
-package uml.diagrams.sequence.sequencediagram;
+package uml.diagrams.sequence.sequencediagrams;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import uml.diagrams.sequence.BaseElement;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
-import uml.diagrams.sequence.lifelines.Lifeline;
-import uml.diagrams.sequence.message.Message;
+import uml.diagrams.sequence.messages.Message;
 
 public class SequenceDiagram extends BaseElement {
 	final static String INVALID_GUARD_CONDITION_VALUE_ERROR_MESSAGE = "A condição de guarda não pode ser vazia ou nula";
