+++ test/uml/diagrams/activity/ActivityDiagramTest.java
@@ -1,0 +1,16 @@
+package uml.diagrams.activity;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class ActivityDiagramTest {
+
+	@BeforeEach
+	void setUp() throws Exception {
+	}
+
+	@Test
+	void test() {
+		new ActivityDiagram();
+	}
+}
