+++ test/uml/diagrams/sequence/fragments/FragmentsToStringTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
-import uml.diagrams.sequence.sequencediagram.SequenceDiagram;
+import uml.diagrams.sequence.sequencediagrams.SequenceDiagram;
 
 public class FragmentsToStringTest {
 	private final static String DEFAULT_NAME = "default";
