+++ test/uml/diagrams/sequence/fragments/OptionalToStringTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
-import uml.diagrams.sequence.sequencediagram.SequenceDiagram;
+import uml.diagrams.sequence.sequencediagrams.SequenceDiagram;
 
 public class OptionalToStringTest {
 	private Optional optional;
