+++ test/uml/diagrams/sequence/messages/MessageAsyncToStringTest.java
@@ -1,4 +1,4 @@
-package uml.diagrams.sequence.message;
+package uml.diagrams.sequence.messages;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -11,22 +11,23 @@ import org.junit.jupiter.params.provider.MethodSource;
 import uml.diagrams.sequence.exceptions.MessageFormatException;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
 import uml.diagrams.sequence.lifelines.Lifeline;
+import uml.diagrams.sequence.messages.MessageAsync;
 
-public class ReplyToStringTest {
-
+public class MessageAsyncToStringTest {
+    
     public static Collection<Object[]> messageParams() {
         return Arrays.asList(new Object[][] {
             {
                 "name", 0.5f, "origem", "destino",
-                "<Reply name=\"name\" prob=\"0.5\" source=\"origem\" target=\"destino\" />"
+                "<MessageAsync name=\"name\" prob=\"0.5\" source=\"origem\" target=\"destino\" />"
             },
             {
                 "name", 1f, "node1", "node2",
-                "<Reply name=\"name\" prob=\"1.0\" source=\"node1\" target=\"node2\" />"
+                "<MessageAsync name=\"name\" prob=\"1.0\" source=\"node1\" target=\"node2\" />"
             },
             {
                 "name", 0f, "node1", "node2",
-                "<Reply name=\"name\" prob=\"0.0\" source=\"node1\" target=\"node2\" />"
+                "<MessageAsync name=\"name\" prob=\"0.0\" source=\"node1\" target=\"node2\" />"
             }
         });
     }
@@ -38,7 +39,7 @@ public class ReplyToStringTest {
                     throws SequenceDiagramRuleException, MessageFormatException {
         Lifeline source = new Lifeline(sourceStr);
         Lifeline target = new Lifeline(targetStr);
-        Reply message = new Reply(name, prob, source, target);
+        MessageAsync message = new MessageAsync(name, prob, source, target);
 
         assertEquals(expectedString, message.toString());
     }
