+++ test/uml/diagrams/sequence/sequencediagrams/SequenceDiagramCreation.java
@@ -1,4 +1,4 @@
-package uml.diagrams.sequence.sequencediagram;
+package uml.diagrams.sequence.sequencediagrams;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -9,6 +9,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
+import uml.diagrams.sequence.sequencediagrams.SequenceDiagram;
 
 
 public class SequenceDiagramCreation {
