+++ test/uml/diagrams/sequence/sequencediagrams/SequenceDiagramEmptyGuardTest.java
@@ -1,4 +1,4 @@
-package uml.diagrams.sequence.sequencediagram;
+package uml.diagrams.sequence.sequencediagrams;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -9,25 +9,24 @@ import java.util.Collection;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import uml.diagrams.sequence.BaseElement;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
+import uml.diagrams.sequence.sequencediagrams.SequenceDiagram;
 
-
-public class SequenceDiagramInvalidNamesTest {
+public class SequenceDiagramEmptyGuardTest {
+    
     public static Collection<Object[]> sequenceDiagramsName() {
         return Arrays.asList(new Object[][] {
-        	{"", true},
-            {null, false}
+            {"name", null }
         });
     }
-
+    
     @ParameterizedTest
     @MethodSource("sequenceDiagramsName")
-    void testErrorCreateSequenceDiagrams(String name, Boolean guardCondition) {
+    void testErrorEmptyGuard(String name, Boolean guardCondition) {
         SequenceDiagramRuleException exception = assertThrows(
                 SequenceDiagramRuleException.class,
                 () -> new SequenceDiagram(name, guardCondition));
         
-        assertEquals(exception.getMessage(), BaseElement.NAME_ERROR);
+        assertEquals(SequenceDiagram.INVALID_GUARD_CONDITION_VALUE_ERROR_MESSAGE , exception.getMessage());
     }
 }
