+++ test/uml/diagrams/sequence/sequencediagrams/SequenceDiagramInvalidNamesTest.java
@@ -1,4 +1,4 @@
-package uml.diagrams.sequence.sequencediagram;
+package uml.diagrams.sequence.sequencediagrams;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -9,23 +9,26 @@ import java.util.Collection;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import uml.diagrams.sequence.BaseElement;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
+import uml.diagrams.sequence.sequencediagrams.SequenceDiagram;
 
-public class SequenceDiagramEmptyGuardTest {
-    
+
+public class SequenceDiagramInvalidNamesTest {
     public static Collection<Object[]> sequenceDiagramsName() {
         return Arrays.asList(new Object[][] {
-            {"name", null }
+        	{"", true},
+            {null, false}
         });
     }
-    
+
     @ParameterizedTest
     @MethodSource("sequenceDiagramsName")
-    void testErrorEmptyGuard(String name, Boolean guardCondition) {
+    void testErrorCreateSequenceDiagrams(String name, Boolean guardCondition) {
         SequenceDiagramRuleException exception = assertThrows(
                 SequenceDiagramRuleException.class,
                 () -> new SequenceDiagram(name, guardCondition));
         
-        assertEquals(SequenceDiagram.INVALID_GUARD_CONDITION_VALUE_ERROR_MESSAGE , exception.getMessage());
+        assertEquals(exception.getMessage(), BaseElement.NAME_ERROR);
     }
 }
